+++ starport/interface/cli/starport/cmd/app.go
@@ -4,8 +4,11 @@ import (
 	"fmt"
 
 	"github.com/spf13/cobra"
+	"github.com/tendermint/starport/starport/pkg/cosmosver"
 	"github.com/tendermint/starport/starport/services/scaffolder"
 )
+
+const sdkVersionFlag = "sdk-version"
 
 // NewApp creates new command named `app` to create Cosmos scaffolds customized
 // by the user given options.
@@ -47,4 +50,17 @@ NOTE: add --verbose flag for verbose (detailed) output.
 `
 	fmt.Printf(message, path)
 	return nil
+}
+
+func addSdkVersionFlag(c *cobra.Command) {
+	c.Flags().String(sdkVersionFlag, string(cosmosver.Launchpad), fmt.Sprintf("Target Cosmos-SDK Version %s", cosmosver.MajorVersions))
+}
+
+func sdkVersion(c *cobra.Command) (cosmosver.MajorVersion, error) {
+	v, _ := c.Flags().GetString(sdkVersionFlag)
+	parsed, err := cosmosver.MajorVersions.Parse(v)
+	if err != nil {
+		return "", fmt.Errorf("%q is an unkown sdk version", v)
+	}
+	return parsed, nil
 }
